This code is to enable states to continue after hitting an assertion.
Currently, given an ASSERT(x), an additional GOTO is generated by the CProver compiler, this is

ASSERT(x)
IF(x) GOTO T;
void(0);
T: SKIP
[... rest of program]

Which in other words triggers a void(0), currently unhanded by Symex (but easily handled by a few lines of code.)
However, after a discussion with Daniel, it may be desirable to maintain states even after a failed assertion.

There are a few ways to do this, e.g. do nothing on void and remerge states, but this generates a pointless GOTO.
For now, this is simply removing GOTOs introduced by assertions.  It may be the compiler can have this option 'turned' off, but this is to be investigated in the future.

Merging code now works on assertions which occur after a previous failure.